requirements.txt
Brotli==1.0.9
click==8.0.3
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.5
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
nltk==3.6.7
numpy==1.22.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
plotly==5.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
six==1.16.0
tenacity==8.0.1
tqdm==4.62.3
Werkzeug==2.0.2
wordcloud==1.8.1
